Add files via upload
The Carpet Calculator app is designed to help users determine the amount of carpet needed to cover a specified area. Users input the height and distance around a room, and the app calculates the required yards of carpet. Additionally, the app features a button that links directly to a carpet purchasing page on Lowe's website, providing a seamless transition from calculating needs to shopping for the product.